+++ bin/folio-list-users.go
@@ -4,7 +4,17 @@ import "os"
 import "fmt"
 import "time"
 import "strconv"
+import "encoding/json"
 import "github.com/indexdata/foliogo"
+
+type user struct {
+	Username string `json:"username"`
+	Active bool `json:"active"`
+}
+type response struct {
+	Users []user `json:"users"`
+	TotalRecords int `json:"totalRecords"y`
+}
 
 func main() {
 	service := foliogo.NewService("https://folio-snapshot-okapi.dev.folio.org")
@@ -19,22 +29,27 @@ func main() {
 		time.Sleep(time.Duration(nsecs) * time.Second)
 	}
 
-	body, err := session.Fetch("users?limit=20", foliogo.RequestParams{})
+	bytes, err := session.Fetch("users?limit=20", foliogo.RequestParams{})
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "%s: fetch users failed: %s\n", os.Args[0], err)
 		os.Exit(2)
 	}
 
-	// Unfortunately, the rendering part is clumsy in Go
-	users := body["users"].([]interface{})
-	for _, e := range users {
-		user := e.(map[string]interface{})
+	var r response
+	err = json.Unmarshal(bytes, &r)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "%s: decode JSON failed: %s\n", os.Args[0], err)
+		os.Exit(2)
+	}
+
+	users := r.Users
+	for _, user := range users {
 		var marker string
-		if user["active"] == true {
+		if user.Active {
 			marker = "*"
 		} else {
 			marker = " "
 		}
-		fmt.Println(marker, user["username"])
+		fmt.Println(marker, user.Username)
 	}
 }