+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change history for @indexdata/foliogo
 
-## [0.0.3](https://github.com/indexdata/foliogo/tree/v0.0.3) (IN PROGRESS)
+## [0.1.0](https://github.com/indexdata/foliogo/tree/v0.1.0) (IN PROGRESS)
 
-* Do not specify an empty body to send when making a GET request. Fixes #7
+* BREAKING: `session.Fetch` returns a byte slice rather than a string-to-any map. Fixes #8.
+* Do not specify an empty body to send when making a GET request. Fixes #7.
 
 ## [0.0.2](https://github.com/indexdata/foliogo/tree/v0.0.2) (2023-11-17)